pkg/profefe: extract from cmd
- move cmd internals to pkg/profefe
- move pkg/profile.Storage to pkg/storage.Storage
- get rid of pkg/profile.Repository
